# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Petshop App",
        "version": "0.1"
    },
    "servers": [
        {
            "url": "/buckApp/public/api/v1"
        }
    ],
    "paths": {
        "/api": {},
        "/admin/login": {
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "User login",
                "description": "Endpoint to authenticate a user.",
                "operationId": "c5a8b2dbaa76d28c359e2dd20c4f599b",
                "requestBody": {
                    "description": "Login credentials",
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "admin@buckhill.co.uk"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "admin"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful login"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/admin/logout": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Log a user out",
                "operationId": "logoutAdminUser",
                "responses": {
                    "200": {
                        "description": "User logged out successfully"
                    },
                    "404": {
                        "description": "User not found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/admin/create": {
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "Create a new user",
                "description": "Create a new user",
                "operationId": "createAdminUser",
                "requestBody": {
                    "description": "User data",
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "$ref": "#/components/schemas/User"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User created successfully"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
        "/admin/user-listing": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "List users",
                "description": "Retrieve a non admin list of users",
                "operationId": "listAdminUsers",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "sortBy",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "desc",
                        "in": "query",
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "first_name",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "phone",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "address",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "created_at",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "500": {
                        "description": "Server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/user-edit/{uuid}": {
            "put": {
                "tags": [
                    "Admin"
                ],
                "summary": "Update a User",
                "description": "Update a user identified by UUID",
                "operationId": "updateAdminUser",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "description": "Uuid of the user to update",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "User data",
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "$ref": "#/components/schemas/User"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User updated successfully"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "User not found"
                    },
                    "500": {
                        "description": "Server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/user-delete/{uuid}": {
            "delete": {
                "tags": [
                    "Admin"
                ],
                "summary": "Delete a user",
                "description": "Deletes a users identified by uuid",
                "operationId": "destroyAdminUser",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "description": "Uuid of the user to delete",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "user deleted successfully"
                    },
                    "404": {
                        "description": "user not found"
                    },
                    "500": {
                        "description": "Server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/user/login": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "User login",
                "description": "Endpoint to authenticate a user.",
                "operationId": "994f48569c4a1bb2132a214f59b24daf",
                "requestBody": {
                    "description": "Login credentials",
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "admin@buckhill.co.uk"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "admin"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful login"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/user/logout": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Log a user out",
                "operationId": "logoutUser",
                "responses": {
                    "200": {
                        "description": "User logged out successfully"
                    },
                    "404": {
                        "description": "User not found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/user/create": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Create a new user",
                "description": "Create a new user",
                "operationId": "createUser",
                "requestBody": {
                    "description": "User data",
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "$ref": "#/components/schemas/User"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
        "/user": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Retrieve a user",
                "description": "Retrieve a user identified by ID",
                "operationId": "getUser",
                "responses": {
                    "200": {
                        "description": "User retrieved successfully"
                    },
                    "404": {
                        "description": "User not found"
                    },
                    "500": {
                        "description": "Server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "User"
                ],
                "summary": "Delete a user",
                "description": "Deletes a users identified by ID",
                "operationId": "destroyUser",
                "responses": {
                    "204": {
                        "description": "user deleted successfully"
                    },
                    "404": {
                        "description": "user not found"
                    },
                    "500": {
                        "description": "Server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/user/edit": {
            "put": {
                "tags": [
                    "User"
                ],
                "summary": "Update a User",
                "description": "Update a user identified by ID",
                "operationId": "updateUser",
                "requestBody": {
                    "description": "User data",
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "$ref": "#/components/schemas/User"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User updated successfully"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "User not found"
                    },
                    "500": {
                        "description": "Server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "User": {
                "title": "User",
                "description": "User resource",
                "required": [
                    "first_name",
                    "last_name",
                    "email",
                    "password",
                    "avatar",
                    "address",
                    "phone_number",
                    "password_confirmation"
                ],
                "properties": {
                    "first_name": {
                        "type": "string",
                        "example": "John"
                    },
                    "last_name": {
                        "type": "string",
                        "example": "Doe"
                    },
                    "email": {
                        "type": "string",
                        "example": "john@example.com"
                    },
                    "password": {
                        "type": "string",
                        "example": "********"
                    },
                    "password_confirmation": {
                        "type": "string",
                        "example": "********"
                    },
                    "avatar": {
                        "type": "string",
                        "nullable": true
                    },
                    "address": {
                        "type": "string",
                        "example": "123 Main St"
                    },
                    "phone_number": {
                        "type": "string",
                        "example": "+123456789"
                    },
                    "is_marketing": {
                        "type": "boolean",
                        "example": false
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}